refactored to functional components & folder structure
