+++ src/client/.env
@@ -1,2 +1,1 @@
 BROWSER=none
-REACT_APP_BACKEND_URL=http://localhost:3000/
