--- src/client/src/App/App.css
@@ -1,30 +1,0 @@
-.App {
-  background-color: #282c34;
-  height: 100vh;
-}
-
-.App-logo {
-  height: 15vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: app-logo-pass-through 10s linear;
-  }
-}
-
-@media (max-width: 500px) {
-  .App-logo {
-    height: 25vmin;
-  }
-}
-
-@keyframes app-logo-pass-through {
-  from {
-    transform: translate(-150%, 0);
-  }
-  to {
-    transform: translate(150%, 0);
-  }
-}
