--- src/client/src/App/Item List/ItemList.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import ExpandableView from './Expandable View/ExpandableView'
-
-class ItemList extends React.Component {
-    render() {
-        return (
-            <ExpandableView></ExpandableView>
-        );
-    }
-}
-
-export default ItemList;
