+++ src/client/src/Components/ExpandableView.css
