+++ src/client/src/Pages/ItemList.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ExpandableView from '../Components/ExpandableView'
+
+const ItemList = () => {
+    return (
+        <ExpandableView></ExpandableView>
+    );
+}
+
+export default ItemList;
