+++ src/client/src/Pages/Login.css
@@ -1,3 +1,4 @@
+
 .login {
     align-items: center;
     color: white;
@@ -7,7 +8,7 @@
     justify-content: center;
     min-height: 100vh;
     text-align: center;
-  }
+}
 
 .login-button {
     background-color: rgb(10, 170, 170);
@@ -19,20 +20,14 @@
     padding: .5rem 1.3rem;
 }
 
-    .login-button:hover {
-        background-color: teal;
-    }
+.login-button:hover {
+    background-color: teal;
+}
 
 .login-container {
     background-color: rgb(117, 117, 117);
     border-radius: 1rem;
     width: 50%;
-}
-
-@media (max-width: 500px) {
-    .login-container {
-        width: 80%;
-    }
 }
 
 .login-content {
@@ -58,7 +53,36 @@
     height: 1.5rem;
 }
 
+.logo {
+    height: 15vmin;
+    pointer-events: none;
+}
+
 .register-link {
     font-size: small;
     margin: .5rem auto;
+}
+
+@media (max-width: 500px) {
+    .login-container {
+        width: 80%;
+    }
+    .logo {
+        height: 25vmin;
+    }
+}
+
+@media (prefers-reduced-motion: no-preference) {
+    .logo {
+        animation: app-logo-pass-through 10s linear;
+    }
+}
+
+@keyframes app-logo-pass-through {
+    from {
+        transform: translate(-150%, 0);
+    }
+    to {
+        transform: translate(150%, 0);
+    }
 }
